Add PersistentConnectionActive and TransientConnectionActive
remove the old [Local/Cluster]ConnectionActive
